requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0
pycryptodome==3.19.0
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
